Rename project Steeltoe.Management.Task to Steeltoe.Management.Tasks, to avoid naming clash with System.Threading.Task.
