+++ src/Management/src/Tasks/PublicAPI.Unshipped.txt
@@ -1,0 +1,12 @@
+#nullable enable
+static Steeltoe.Management.Tasks.TaskServiceCollectionExtensions.AddTask(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services, Steeltoe.Common.IApplicationTask! task) -> void
+static Steeltoe.Management.Tasks.TaskServiceCollectionExtensions.AddTask(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services, string! name, System.Action<System.IServiceProvider!>! runAction) -> void
+static Steeltoe.Management.Tasks.TaskServiceCollectionExtensions.AddTask(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services, string! name, System.Action<System.IServiceProvider!>! runAction, Microsoft.Extensions.DependencyInjection.ServiceLifetime lifetime) -> void
+static Steeltoe.Management.Tasks.TaskServiceCollectionExtensions.AddTask(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services, System.Func<System.IServiceProvider!, Steeltoe.Common.IApplicationTask!>! factory) -> void
+static Steeltoe.Management.Tasks.TaskServiceCollectionExtensions.AddTask(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services, System.Func<System.IServiceProvider!, Steeltoe.Common.IApplicationTask!>! factory, Microsoft.Extensions.DependencyInjection.ServiceLifetime lifetime) -> void
+static Steeltoe.Management.Tasks.TaskServiceCollectionExtensions.AddTask<T>(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services) -> void
+static Steeltoe.Management.Tasks.TaskServiceCollectionExtensions.AddTask<T>(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services, Microsoft.Extensions.DependencyInjection.ServiceLifetime lifetime) -> void
+static Steeltoe.Management.Tasks.TaskWebHostExtensions.RunWithTasks(this Microsoft.AspNetCore.Hosting.IWebHost! webHost) -> void
+static Steeltoe.Management.Tasks.TaskWebHostExtensions.RunWithTasks(this Microsoft.Extensions.Hosting.IHost! host) -> void
+Steeltoe.Management.Tasks.TaskServiceCollectionExtensions
+Steeltoe.Management.Tasks.TaskWebHostExtensions
