+++ src/Management/src/Tasks/Steeltoe.Management.Tasks.csproj
