+++ src/Management/src/Tasks/TaskServiceCollectionExtensions.cs
@@ -5,7 +5,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using Steeltoe.Common;
 
-namespace Steeltoe.Management.Task;
+namespace Steeltoe.Management.Tasks;
 
 public static class TaskServiceCollectionExtensions
 {
