+++ src/Steeltoe.Management.slnf
@@ -16,12 +16,12 @@
       "Management\\src\\Abstractions\\Steeltoe.Management.Abstractions.csproj",
       "Management\\src\\Endpoint\\Steeltoe.Management.Endpoint.csproj",
       "Management\\src\\Prometheus\\Steeltoe.Management.Prometheus.csproj",
-      "Management\\src\\Task\\Steeltoe.Management.Task.csproj",
+      "Management\\src\\Tasks\\Steeltoe.Management.Tasks.csproj",
       "Management\\src\\Tracing\\Steeltoe.Management.Tracing.csproj",
       "Management\\src\\Wavefront\\Steeltoe.Management.Wavefront.csproj",
       "Management\\test\\Endpoint.Test\\Steeltoe.Management.Endpoint.Test.csproj",
       "Management\\test\\Prometheus.Test\\Steeltoe.Management.Prometheus.Test.csproj",
-      "Management\\test\\Task.Test\\Steeltoe.Management.Task.Test.csproj",
+      "Management\\test\\Tasks.Test\\Steeltoe.Management.Tasks.Test.csproj",
       "Management\\test\\Tracing.Test\\Steeltoe.Management.Tracing.Test.csproj",
       "Management\\test\\Wavefront.Test\\Steeltoe.Management.Wavefront.Test.csproj",
       "Tools\\src\\ConfigurationSchemaGenerator\\ConfigurationSchemaGenerator.csproj"
